ui-community can modify team members
