+++ ui/ui-community/pages/team/initialize.tsx
@@ -1,0 +1,1 @@
+export { default } from "@grouparoo/ui-components/pages/team/initialize";
