+++ ui/ui-components/pages/team/[id]/teamMember/new.tsx
@@ -1,0 +1,1 @@
+export { default } from "../../../teamMember/new";
