+++ ui/ui-enterprise/pages/team/[id]/teamMember/new.tsx
@@ -1,1 +1,1 @@
-export { default } from "../../../teamMember/new";
+export { default } from "@grouparoo/ui-components/pages/team/[id]/teamMember/new";
