+++ ui/ui-enterprise/pages/team/initiailize.tsx
@@ -1,0 +1,1 @@
+export { default } from "@grouparoo/ui-components/pages/team/initialize";
